Parking v1 complete

+++ src/Models/Payment/CashPayment.java
@@ -3,7 +3,7 @@ package Models.Payment;
 public class CashPayment implements PaymentStratergy {
 
     @Override
-    public void processPayment(int amount) {
+    public void processPayment(double amount) {
         System.out.println("Payment processed using Cash for amount: "+amount);
     }
