+++ src/Models/Payment/CreditCardPayment.java
@@ -2,7 +2,7 @@ package Models.Payment;
 
 public class CreditCardPayment implements PaymentStratergy{
     @Override
-    public void processPayment(int amount) {
+    public void processPayment(double amount) {
         System.out.println("Payment processed using Credit Card for amount: "+amount);
     }
 }
