+++ src/Models/Payment/PaymentStratergy.java
@@ -1,5 +1,5 @@
 package Models.Payment;
 
 public interface PaymentStratergy {
-    public void processPayment(int amount);
+    public void processPayment(double amount);
 }
