+++ src/Models/Payment/UPIPayment.java
@@ -2,7 +2,7 @@ package Models.Payment;
 
 public class UPIPayment implements PaymentStratergy {
     @Override
-    public void processPayment(int amount) {
+    public void processPayment(double amount) {
         System.out.println("Payment processed using UPI for amount: "+amount);
     }
 }
